get proper fonts for item popup
